move guest Ray type to shapes.rs
